Extract ToolSelectionDocument operations into document_config/tool_selection.rs
